+++ components/admin_console/billing/billing_subscriptions.scss
@@ -73,16 +73,7 @@
 }
 
 .UpgradeMattermostCloud {
-    padding: 32px;
-    display: flex;
-    flex-direction: column;
-    width: 332px;
-    margin-left: 20px;
-    align-items: center;
-    background-color: var(--sys-center-channel-bg);
-    border: 1px solid rgba(var(--sys-center-channel-color-rgb), 0.08);
-    box-shadow: 0px 2px 3px rgba(0, 0, 0, 0.08);
-    border-radius: 4px;
+    text-align: center;
 }
 
 .UpgradeMattermostCloud__image {
