+++ components/admin_console/billing/billing_subscriptions.tsx
@@ -3,7 +3,6 @@
 
 import React, {useEffect, useState} from 'react';
 import {useDispatch, useStore, useSelector} from 'react-redux';
-import {FormattedMessage, useIntl} from 'react-intl';
 
 import {getStandardAnalytics} from 'mattermost-redux/actions/admin';
 import {getCloudSubscription, getCloudProducts, getCloudCustomer} from 'mattermost-redux/actions/cloud';
@@ -15,27 +14,26 @@ import {DispatchFunc} from 'mattermost-redux/types/actions';
 import {PreferenceType} from 'mattermost-redux/types/preferences';
 
 import {pageVisited, trackEvent} from 'actions/telemetry_actions';
-import {openModal} from 'actions/views/modals';
-import AlertBanner from 'components/alert_banner';
-import BlockableLink from 'components/admin_console/blockable_link';
-import FormattedMarkdownMessage from 'components/formatted_markdown_message';
-import PurchaseModal from 'components/purchase_modal';
 import FormattedAdminHeader from 'components/widgets/admin_console/formatted_admin_header';
 import {getCloudContactUsLink, InquiryType, InquiryIssue} from 'selectors/cloud';
 import {GlobalState} from 'types/store';
 import {
     Preferences,
     CloudBanners,
-    ModalIdentifiers,
     TELEMETRY_CATEGORIES,
 } from 'utils/constants';
 import {isCustomerCardExpired} from 'utils/cloud_utils';
 
-import privateCloudImage from 'images/private-cloud-image.svg';
-import upgradeMattermostCloudImage from 'images/upgrade-mattermost-cloud-image.svg';
-
 import BillingSummary from './billing_summary';
 import PlanDetails from './plan_details';
+
+import {
+    contactSalesCard,
+    cancelSubscription,
+    infoBanner,
+    creditCardExpiredBanner,
+    paymentFailedBanner,
+} from './billing_subscriptions_jsx_pieces';
 
 import './billing_subscriptions.scss';
 
@@ -45,7 +43,6 @@ type Props = {
 };
 
 const BillingSubscriptions: React.FC<Props> = () => {
-    const {formatMessage} = useIntl();
     const dispatch = useDispatch<DispatchFunc>();
     const store = useStore();
     const userLimit = useSelector((state: GlobalState) => parseInt(getConfig(state).ExperimentalCloudUserLimit!, 10));
@@ -64,14 +61,17 @@ const BillingSubscriptions: React.FC<Props> = () => {
 
     const [showCreditCardBanner, setShowCreditCardBanner] = useState(true);
 
-    const onUpgradeMattermostCloud = () => {
-        trackEvent('cloud_admin', 'click_upgrade_mattermost_cloud');
+    // TODO: replace this temporal hardcoded value with the right state value
+    enum SubscriptionType {
+        FREE_TRIAL = 'FREE_TRIAL',
+        CLOUD_STARTER = 'CLOUD_STARTER',
+        CLOUD_PROFESSIONAL = 'CLOUD_PROFESSIONAL',
+        CLOUD_ENTERPRISE = 'CLOUD_ENTERPRISE'
+    }
 
-        dispatch(openModal({
-            modalId: ModalIdentifiers.CLOUD_PURCHASE,
-            dialogType: PurchaseModal,
-        }));
-    };
+    // TODO: remove these hardcoded values
+    const typeSubscription = SubscriptionType.FREE_TRIAL;
+    const daysLeft = 13;
 
     useEffect(() => {
         getCloudSubscription()(dispatch, store.getState());
@@ -122,104 +122,11 @@ const BillingSubscriptions: React.FC<Props> = () => {
         ]));
     };
 
-    const upgradeMattermostCloud = () => (
-        <div className='UpgradeMattermostCloud'>
-            <div className='UpgradeMattermostCloud__image'>
-                <img src={upgradeMattermostCloudImage}/>
-            </div>
-            <div className='UpgradeMattermostCloud__title'>
-                <FormattedMessage
-                    id='admin.billing.subscription.upgradeMattermostCloud.title'
-                    defaultMessage='Need more users?'
-                />
-            </div>
-            <div className='UpgradeMattermostCloud__description'>
-                <FormattedMarkdownMessage
-                    id='admin.billing.subscription.upgradeMattermostCloud.description'
-                    defaultMessage='The free tier is **limited to 10 users.** Get access to more users, teams and other great features'
-                />
-            </div>
-            <button
-                type='button'
-                onClick={onUpgradeMattermostCloud}
-                className='UpgradeMattermostCloud__upgradeButton'
-            >
-                <FormattedMessage
-                    id='admin.billing.subscription.upgradeMattermostCloud.upgradeButton'
-                    defaultMessage='Upgrade Mattermost Cloud'
-                />
-            </button>
-        </div>
-    );
-
-    const privateCloudCard = () => (
-        <div className='PrivateCloudCard'>
-            <div className='PrivateCloudCard__text'>
-                <div className='PrivateCloudCard__text-title'>
-                    <FormattedMessage
-                        id='admin.billing.subscription.privateCloudCard.title'
-                        defaultMessage='Looking for a high-trust private cloud?'
-                    />
-                </div>
-                <div className='PrivateCloudCard__text-description'>
-                    <FormattedMessage
-                        id='admin.billing.subscription.privateCloudCard.description'
-                        defaultMessage='If you need software with dedicated, single-tenant architecture, Mattermost Private Cloud (Beta) is the solution for high-trust collaboration.'
-                    />
-                </div>
-                <a
-                    href={contactSalesLink}
-                    rel='noopener noreferrer'
-                    target='_new'
-                    className='PrivateCloudCard__contactSales'
-                    onClick={() => trackEvent('cloud_admin', 'click_contact_sales')}
-                >
-                    <FormattedMessage
-                        id='admin.billing.subscription.privateCloudCard.contactSales'
-                        defaultMessage='Contact Sales'
-                    />
-                </a>
-            </div>
-            <div className='PrivateCloudCard__image'>
-                <img src={privateCloudImage}/>
-            </div>
-        </div>
-    );
-
-    const cancelSubscription = () => (
-        <div className='cancelSubscriptionSection'>
-            <div className='cancelSubscriptionSection__text'>
-                <div className='cancelSubscriptionSection__text-title'>
-                    <FormattedMessage
-                        id='admin.billing.subscription.cancelSubscriptionSection.title'
-                        defaultMessage='Cancel your subscription'
-                    />
-                </div>
-                <div className='cancelSubscriptionSection__text-description'>
-                    <FormattedMessage
-                        id='admin.billing.subscription.cancelSubscriptionSection.description'
-                        defaultMessage='At this time, deleting a workspace can only be done with the help of a customer support representative.'
-                    />
-                </div>
-                <a
-                    href={cancelAccountLink}
-                    rel='noopener noreferrer'
-                    target='_new'
-                    className='cancelSubscriptionSection__contactUs'
-                    onClick={() => trackEvent('cloud_admin', 'click_contact_us')}
-                >
-                    <FormattedMessage
-                        id='admin.billing.subscription.cancelSubscriptionSection.contactUs'
-                        defaultMessage='Contact Us'
-                    />
-                </a>
-            </div>
-        </div>
-    );
-
     if (!subscription || !products) {
         return null;
     }
+
+    const isPaidTier= Boolean(subscription?.is_paid_tier === 'true');
 
     return (
         <div className='wrapper--fixed BillingSubscriptions'>
@@ -229,88 +136,22 @@ const BillingSubscriptions: React.FC<Props> = () => {
             />
             <div className='admin-console__wrapper'>
                 <div className='admin-console__content'>
-                    {shouldShowPaymentFailedBanner() && (
-                        <AlertBanner
-                            mode='danger'
-                            title={formatMessage({
-                                id: 'billing.subscription.info.mostRecentPaymentFailed',
-                                defaultMessage: 'Your most recent payment failed',
-                            })}
-                            message={
-                                <>
-                                    <FormattedMessage
-                                        id='billing.subscription.info.mostRecentPaymentFailed.description.mostRecentPaymentFailed'
-                                        defaultMessage='It looks your most recent payment failed because the credit card on your account has expired. Please '
-                                    />
-                                    <BlockableLink
-                                        to='/admin_console/billing/payment_info'
-                                    >
-                                        <FormattedMessage
-                                            id='billing.subscription.info.mostRecentPaymentFailed.description.updatePaymentInformation'
-                                            defaultMessage='update your payment information'
-                                        />
-                                    </BlockableLink>
-                                    <FormattedMessage
-                                        id='billing.subscription.info.mostRecentPaymentFailed.description.avoidAnyDisruption'
-                                        defaultMessage=' to avoid any disruption.'
-                                    />
-                                </>
-                            }
-                        />
-                    )}
-                    {shouldShowInfoBanner() && (
-                        <AlertBanner
-                            mode='info'
-                            title={formatMessage({
-                                id: 'billing.subscription.info.headsup',
-                                defaultMessage: 'Just a heads up',
-                            })}
-                            message={formatMessage({
-                                id: 'billing.subscription.info.headsup.description',
-                                defaultMessage:
-                    'You’re nearing the user limit with the free tier of Mattermost Cloud. We’ll let you know if you hit that limit.',
-                            })}
-                            onDismiss={() => handleHide()}
-                        />
-                    )}
-                    {showCreditCardBanner && isCardExpired && (
-                        <AlertBanner
-                            mode='danger'
-                            title={
-                                <FormattedMessage
-                                    id='admin.billing.subscription.creditCardHasExpired'
-                                    defaultMessage='Your credit card has expired'
-                                />
-                            }
-                            message={
-                                <>
-                                    <FormattedMessage
-                                        id='admin.billing.subscription.creditCardHasExpired.please'
-                                        defaultMessage='Please '
-                                    />
-                                    <BlockableLink
-                                        to='/admin_console/billing/payment_info'
-                                    >
-                                        <FormattedMessage
-                                            id='admin.billing.subscription.creditCardHasExpired.description.updatePaymentInformation'
-                                            defaultMessage='update your payment information'
-                                        />
-                                    </BlockableLink>
-                                    <FormattedMessage
-                                        id='admin.billing.subscription.creditCardHasExpired.description.avoidAnyDisruption'
-                                        defaultMessage=' to avoid any disruption.'
-                                    />
-                                </>
-                            }
-                            onDismiss={() => setShowCreditCardBanner(false)}
-                        />
-                    )}
+                    {shouldShowPaymentFailedBanner() && paymentFailedBanner()}
+                    {shouldShowInfoBanner() && infoBanner(handleHide)}
+                    {showCreditCardBanner && isCardExpired && creditCardExpiredBanner(setShowCreditCardBanner)}
                     <div className='BillingSubscriptions__topWrapper'>
-                        <PlanDetails/>
-                        {subscription?.is_paid_tier === 'true' ? <BillingSummary/> : upgradeMattermostCloud()}
+                        <PlanDetails
+                            typeSubscription={typeSubscription}
+                            daysLeft={13}
+                        />
+                        <BillingSummary
+                            isPaidTier={isPaidTier}
+                            typeSubscription={typeSubscription}
+                            daysLeft={13}
+                        />
                     </div>
-                    {privateCloudCard()}
-                    {cancelSubscription()}
+                    {contactSalesCard(contactSalesLink, typeSubscription, isPaidTier)}
+                    {cancelSubscription(cancelAccountLink, typeSubscription, isPaidTier)}
                 </div>
             </div>
         </div>
