+++ i18n/en.json
@@ -291,7 +291,7 @@
   "admin.billing.subscription.planDetails.howBillingWorks": "See how billing works",
   "admin.billing.subscription.planDetails.numberOfSeats": "{numberOfSeats} seats",
   "admin.billing.subscription.planDetails.numberOfSeatsRegistered": "({userCount} currently registered)",
-  "admin.billing.subscription.planDetails.perUserPerMonth": "/user/month",
+  "admin.billing.subscription.planDetails.perUserPerMonth": "/user/month.",
   "admin.billing.subscription.planDetails.planDetailsName.freeForXOrMoreUsers": "/user/month for {aboveUserLimit} or more users.",
   "admin.billing.subscription.planDetails.planDetailsName.freeUpTo": "Free for up to {aboveUserLimit} users.",
   "admin.billing.subscription.planDetails.prolongedOverages": "Prolonged overages may result in additional charges.",
